root-page-cat.jsp update problem fixed
change catcontroller path by globalProjectImgPath
